lambda/requirements.txt
python-dotenv
google_auth_oauthlib
google-api-python-client
google-api-core
openai